improve routing arg conversion
